system/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
cloudinary~=1.44.0
dotenv~=0.9.9
python-dotenv~=1.1.0
requests~=2.32.3